Setting WSGI_MIDDLEWARE_FUNCTIONS created
What it does:

This setting allows one to wire WSGI middlewares to Django easily.
Although is already possible to do so, there is no default way to do so.

Is reasonable to think that new deployments should be done by WSGI, as is
the recomended way, so I feel not bad about non-wsgi-based options not
taking this in consideration. But is worth to put some note on docs saying
that this setting is a no-op in such cases.

My reason to code it is to allow easily interchance between Django and
other Python frameworks.

Usage:

- It defines the order of WSGI functions, starting from later ones on response
as Django middlewares does.

- It applies only to deploy methods based on WSGIHandler, as FastCGI and so.

- If this setting is not defined, use only default WSGIWorker.

- For now, only Python callables are ok as list itens, but I plan to allow
string path for modules too, as Django middleware tuple allows.